requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
fastapi==0.135.1
fastembed==0.7.4
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.13
langchain-core==1.2.21
langchain-ollama==1.0.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
numpy==2.4.3
ollama==0.6.1
onnxruntime==1.24.4
orjson==3.11.7
ormsgpack==1.12.2
packaging @ file:///home/task_176104885106445/conda-bld/packaging_1761049078006/work
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==11.3.0
pluggy==1.6.0
portalocker==3.2.0
protobuf==7.34.1
py_rust_stemmers==0.1.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
pypdfium2==5.6.0
pytest==9.0.2
python-docx==1.2.0
python-dotenv==1.2.2
PyYAML==6.0.3
qdrant-client==1.17.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
shellingham==1.5.4
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
xxhash==3.6.0
zstandard==0.25.0
